done find

+++ angular-training/src/app/pages/quanlythuvien/bandoc/bandoc.component.ts
@@ -3,6 +3,7 @@ import { Validators, FormBuilder } from '@angular/forms';
 import { ToastrService } from 'ngx-toastr';
 import { Page } from 'src/app/shared/model/page';
 import { BandocService } from 'src/app/shared/services/bandoc.service';
+import Swal from "sweetalert2";
 
 @Component({
     selector: 'app-bandoc',
@@ -97,18 +98,48 @@ export class BandocComponent implements OnInit {
         this.formGroup.patchValue(row);
     }
 
+    // delete(id: any) {
+    //     this.BandocService.deleteBandoc(id).subscribe({
+    //         next: () => {
+    //             this.toastrService.success(`Successful`);
+    //             this.formGroup.reset();
+    //             this.getBandoc();
+    //         },
+    //         error: (error) => {
+    //             this.toastrService.error(`Failed !!!`);
+    //             console.error(error);
+    //         },
+    //     });
+    // }
     delete(id: any) {
-        this.BandocService.deleteBandoc(id).subscribe({
-            next: () => {
-                this.toastrService.success(`Successful`);
-                this.formGroup.reset();
-                this.getBandoc();
-            },
-            error: (error) => {
-                this.toastrService.error(`Failed !!!`);
-                console.error(error);
-            },
-        });
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                this.BandocService.deleteBandoc(id).subscribe({
+                    next: () => {
+                        this.toastrService.success('Successful');
+                        this.formGroup.reset();
+                        this.getBandoc();
+                    },
+                    error: (error) => {
+                        this.toastrService.error('Failed !!!');
+                        console.error(error);
+                    },
+                });
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                )
+            }
+        })
     }
     resetForm() {
         this.formGroup.reset();
