+++ angular-training/src/app/pages/quanlythuvien/nhaxuatban/nhaxuatban.component.ts
@@ -3,6 +3,7 @@ import { Validators, FormBuilder } from '@angular/forms';
 import { ToastrService } from 'ngx-toastr';
 import { Page } from 'src/app/shared/model/page';
 import { NhaXuatBanService } from 'src/app/shared/services/nhaxuatban.service';
+import Swal from "sweetalert2";
 @Component({
   selector: 'app-nhaxuatban',
   templateUrl: './nhaxuatban.component.html',
@@ -93,18 +94,36 @@ export class NhaxuatbanComponent implements OnInit {
         this.formGroup.patchValue(row);
     }
 
+
     delete(id: any) {
-        this.NhaXuatBanService.deleteNhaXuatBan(id).subscribe({
-            next: () => {
-                this.toastrService.success(`Successful`);
-                this.formGroup.reset();
-                this.getNhaXuatBan();
-            },
-            error: (error) => {
-                this.toastrService.error(`Failed !!!`);
-                console.error(error);
-            },
-        });
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                this.NhaXuatBanService.deleteNhaXuatBan(id).subscribe({
+                    next: () => {
+                        this.toastrService.success('Successful');
+                        this.formGroup.reset();
+                        this.getNhaXuatBan();
+                    },
+                    error: (error) => {
+                        this.toastrService.error('Failed !!!');
+                        console.error(error);
+                    },
+                });
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                )
+            }
+        })
     }
     resetForm() {
         this.formGroup.reset();
