+++ angular-training/src/app/pages/quanlythuvien/sach/sach.component.ts
@@ -5,6 +5,7 @@ import { Page } from 'src/app/shared/model/page';
 import { SachService } from 'src/app/shared/services/sach.service';
 import { NhaXuatBanService } from 'src/app/shared/services/nhaxuatban.service';
 import { TacGiaService } from 'src/app/shared/services/tacgia.service';
+import Swal from "sweetalert2";
 @Component({
     selector: 'app-sach',
     templateUrl: './sach.component.html',
@@ -132,18 +133,36 @@ export class SachComponent implements OnInit {
 
         });
     }
+
     delete(id: any) {
-        this.SachService.deleteSach(id).subscribe({
-            next: () => {
-                this.toastrService.success(`Successful`);
-                this.formGroup.reset();
-                this.getSach();
-            },
-            error: (error) => {
-                this.toastrService.error(`Failed !!!`);
-                console.error(error);
-            },
-        });
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                this.SachService.deleteSach(id).subscribe({
+                    next: () => {
+                        this.toastrService.success('Successful');
+                        this.formGroup.reset();
+                        this.getSach();
+                    },
+                    error: (error) => {
+                        this.toastrService.error('Failed !!!');
+                        console.error(error);
+                    },
+                });
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                )
+            }
+        })
     }
     resetForm() {
         this.formGroup.reset();
