+++ angular-training/src/app/pages/quanlythuvien/tacgia/tacgia.component.ts
@@ -3,6 +3,7 @@ import { Validators, FormBuilder } from '@angular/forms';
 import { ToastrService } from 'ngx-toastr';
 import { Page } from 'src/app/shared/model/page';
 import { TacGiaService } from 'src/app/shared/services/tacgia.service';
+import Swal from "sweetalert2";
 @Component({
   selector: 'app-tacgia',
   templateUrl: './tacgia.component.html',
@@ -29,7 +30,7 @@ export class TacgiaComponent implements OnInit {
         diaChi: ['', Validators.maxLength(100)],
         sdt: [''],
         moTa: [''],
-      
+
     });
 
     constructor(
@@ -93,18 +94,36 @@ export class TacgiaComponent implements OnInit {
         this.formGroup.patchValue(row);
     }
 
+
     delete(id: any) {
-        this.TacGiaService.deleteTacGia(id).subscribe({
-            next: () => {
-                this.toastrService.success(`Successful`);
-                this.formGroup.reset();
-                this.getTacGia();
-            },
-            error: (error) => {
-                this.toastrService.error(`Failed !!!`);
-                console.error(error);
-            },
-        });
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                this.TacGiaService.deleteTacGia(id).subscribe({
+                    next: () => {
+                        this.toastrService.success('Successful');
+                        this.formGroup.reset();
+                        this.getTacGia();
+                    },
+                    error: (error) => {
+                        this.toastrService.error('Failed !!!');
+                        console.error(error);
+                    },
+                });
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                )
+            }
+        })
     }
     resetForm() {
         this.formGroup.reset();
